Revert "Weekly dev merge (#34)"
This reverts commit 3bb55e0c7881399b9eac9f99fb132f3394093526.
